v3.0.1: fixed sniper's mark % damage bonus being loaded as an int

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/buffs/SnipersMark.java
@@ -75,7 +75,7 @@ public class SnipersMark extends FlavourBuff implements ActionIndicator.Action {
 	public void restoreFromBundle( Bundle bundle ) {
 		super.restoreFromBundle( bundle );
 		object = bundle.getInt( OBJECT );
-		percentDmgBonus = bundle.getInt( BONUS );
+		percentDmgBonus = bundle.getFloat( BONUS );
 	}
 
 	@Override
